Implemented Indentation
Logical Expressions with parens are a mess. It's imo a bug in the parser: It should report '(a)' as an expression, but it does not. Only 'a' is reported, so no indention level can be derived.

'else if' also still needs some special treatment. Else it looks alright. Probably needs a lot more testing. At least I am confident that it will not alter the functionality of the formatted program code.